refactor: di 패키지의 위치를 최상단으로 변경
